[CON-45 NFID] Auto fill in wallet principal ID after connecting wallet
* for New Beam (Get Paid) and New Meeting
